Share ferment lifecycle helpers
Remove the dangling transition-doc link and centralize phase terminal settling plus single-phase activation cleanup so state transitions, event replay, and legacy storage use one implementation.
